[CL-2210] Move image cropper
